Fix parsing, add option to emit table
This fixes parsing of some LLVM IR instructions.

The cargo wrapper is also updated, such that it also catches other LLVM
IR files and object files. This means that when a project is also
building C files in a build.rs file with the `-fstack-size-section`
flag, the stack size of the C builds are also used for calculating the
stack sizes. However, I'm not sure how accurate this is.

I also added the option to emit a table in the terminal instead of
generating a dot file. Dot files can be very big for bigger project, and
thus converting the dot file to an SVG could be impossible or take a
very long time.